\name{AMS_inv}
\alias{AMS_inv}
\title{
Inverse eigenvectors calculator
}
\description{
Function that generate an eigenvectors matrix from inverse anisotropy matrix.
}
\usage{
AMS_inv(mat, type="v")
}
\arguments{
 \item{mat}{matrix to be inverted}
 \item{type}{character string giving the type of input matrix.

 "v"= V1, V1_dec, V1_inc, V2, V2_dec, V2_inc, V3, V3_dec, V3_inc.

 "m"= K11,K22,K33,K12,K23,K13.

 "m3x3"= take the 3x3 matrix generated by PmagDiR::matrix_maker}

}
\details{
AMS measurements are often given as "m" type (=K11,K22,K33,K12,K23,K13). Depending on the data available, different matrix can be upload, including matrix (3x3) generated by the matrix_maker function.
}
\value{
  \item{L}{Lineation (K1/K2) of the original matrix}
  \item{F}{Foliation (K2/K3) of the original matrix}
  \item{P}{Anisotropy degree (K1/K3) of the original matrix}
The output file is the 3x3 matrix of the inverted matrix eigenvectors.
}
\references{
}
\author{
Edoardo Dallanave
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##Generate triaxial matrix with K1=North,K2=Est,K3= down, return matrix and anisotropy degree P.
T = matrix_maker(Fol=2,Lin=1,v1d=0,v1i=0,v2d=90,v2i=0,v3d=0,v3i=90)
T
##Inverts eigenvectors matrix
Inv = AMS_inv(T,type="m3x3")
Inv
}
