% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Ardo_Geo_PmagDiR}
\alias{Ardo_Geo_PmagDiR}
\title{Ardo section geographic directions and bedding tilt}
\format{
## `Ardo_Geo_PmagDiR`
A data frame with 239 rows and 5 columns:
\describe{
  \item{G_dec}{Paleomagnetic declination}
  \item{G_inc}{Paleomagnetic inclination}
  \item{B_az}{Bedding dip azimuth}
  \item{B_plung}{Bedding Plunge}
  \item{Position}{Stratigraphic position}

  ...
}
}
\source{
<https://www.sciencedirect.com/science/article/abs/pii/S0031018212002027>
}
\usage{
Ardo_Geo_PmagDiR
}
\description{
Paleomagnetic declination and incliantion, in geographic coordinates, and bedding azimuth and plunge, from the Paleogene South
Ardo section of the Venetian Southern Alps (Italy)
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
