% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Ardo_PmagDirs}
\alias{Ardo_PmagDirs}
\title{Ardo section paleomagnetic directions}
\format{
## `Ardo_PmagDirs`
A data frame with 239 rows and 2 columns:
\describe{
  \item{B.Dec}{Paleomagnetic declination}
  \item{B.Inc}{Paleomagnetic inclination}
  ...
}
}
\source{
<https://www.sciencedirect.com/science/article/abs/pii/S0031018212002027>
}
\usage{
Ardo_PmagDirs
}
\description{
Paleomagnetic declination and incliantion, in tilt-correted coordinates, from the Paleogene South
Ardo section of the Venetian Southern Alps (Italy)
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
