\name{curve_cross}
\alias{curve_cross}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Crossing point
}
\description{
Finds crossing points of two empirical curves
}
\usage{
curve_cross(a,b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{a}{first series of x-y data points}
  \item{b}{second series of x-y data points}
}
\details{
This function is used by other functions in the PmagDiR package
}
\value{
Return a (x,y) file with the crossing point of the two input (empirical) curves
}
\references{
}
\author{
Edoardo Dallanave
}
\note{
}


\seealso{
}
\examples{
}
