\name{equalarea}
\alias{equalarea}
\title{
Equal area
}
\description{
Draw an empty equal area plot with title option.
}
\usage{
equalarea(title="")
}
\arguments{
  \item{title}{string added below the equal area plot}
}
\details{
This function is used by several other functions for plotting directions
}
\value{
}
\references{
}
\author{
Edoardo Dallanave
}
\note{
}

\seealso{
}
\examples{
equalarea(title="equal area")
}
