% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{km_AMS}
\alias{km_AMS}
\title{Koumac section anisotropy of magnetic suceptibility}
\format{
## `km_AMS`
A data frame with 1 row and 9 columns:
\describe{
  \item{V1}{max eigenvalue}
  \item{V1_dec}{max eigenvalue declination}
  \item{V1_inc}{max eigenvalue inclination}
  \item{V2}{intermediate eigenvalue}
  \item{V2_dec}{intermediate eigenvalue declination}
  \item{V2_inc}{intermediate eigenvalue inclination}
  \item{V3}{min eigenvalue}
  \item{V3_dec}{min eigenvalue declination}
  \item{V3_inc}{min eigenvalue inclination}
  ...
}
}
\source{
<https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2019GC008699>
}
\usage{
km_AMS
}
\description{
Eigenvalues and eigenvectors of the average AMS from the Eocene
Koumac section of New Caledonia
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
