% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{km_PmagDirs}
\alias{km_PmagDirs}
\title{Koumac section geographic paleomagnetic directions and bedding}
\format{
## `km_PmagDirs`
A data frame with 88 rows and 4 columns:
\describe{
  \item{Dec}{Paleomagnetic declination}
  \item{Inc}{Paleomagnetic inclination}
  \item{Baz}{Bedding dip azimuth}
  \item{Bdip}{Bedding dip azimuth}
  ...
}
}
\source{
<https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2019GC008699>
}
\usage{
km_PmagDirs
}
\description{
Paleomagnetic declination and incliantion, in geographic coordinates, and bedding azimuth and plunge, from the Eocene
Koumac section of New Caledonia
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
