\name{PPCA_HR16}
\alias{PPCA_HR16}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
PPCA analysis: Anchor or not to anchor?
}
\description{
Perform Bayesian model selection for non-anchoring, anchoring, or constraining a paleomagnetic direction
}
\usage{
PPCA_HR16(VEPs)
}
\arguments{
  \item{VEPs}{Vector end-points of demagnetization steps in cartesian [x,y,z] coordinates}
}
\details{
Return a list of the three possible interpolations and the parameters as described in Heslop and Roberts (2016), with recomendation of the best interpolation.
}
\value{
}
\references{
Heslop, D., & Roberts, A. P. (2016). Analyzing paleomagnetic data: To anchor or not to anchor? Journal of Geophysical Research: Solid Earth, 121(11), 7742–7753. https://doi.org/10.1002/2016JB013387
}
\author{
Edoardo Dallanave}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

