\name{save_pdf}
\alias{save_pdf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Save pdf figure
}
\description{
Function that uses the dev.print function to save figure in a pre-determinded size pdf file within the working directory.
}
\usage{
save_pdf(name="Figure.pdf",width=11,height=8)
}
\arguments{
  \item{name}{name of the exported figure. Default is "Figure.pdf"}
  \item{width}{width of the exported figure. Default is 11 inches}
  \item{height}{height of the exported figure. Default is 8 inches}
}

\details{
}
\value{
}
\references{
}
\author{
Edoardo Dallanave
}
\note{
}


\seealso{
}
\examples{
}
