\name{strain_DI}
\alias{strain_DI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Paleomagnetic directions strainer
}
\description{
Function that deviates paleomagnetic directions from the original orientation by applying a deformation (strain) matrix.
}
\usage{
strain_DI(DIAP,M,export=FALSE,name="strained_dirs")
}
\arguments{
  \item{DIAP}{declination and inclination,bedding-dip azimuth and plunge as single file with four colums}
  \item{M}{strain matrix}
  \item{export}{If TRUE, results are automatically exported as .csv file in the working folder. Default is FALSE}
  \item{name}{name of the exported file. Default is "strained_dirs"}
}
\details{
The strain matrix is a 3x3 (k1+k2+k3=3) symmetric matrix like the one generated by PmagDirs::matrix_maker. It simulates the effect of strain on the paleomagnetic singal and bedding
}
\value{
The function returns a data frame with four colums containing strained declination, inclination, bedding-dip azimuth and plunge, like the example produced in Dallanave and Kirscher (2020).
}
\references{
Dallanave, E. & Kirscher, U., Front. Earth Sci. Geomagn. Paleomagn. 8:592277, 1–16 (2020).
}
\author{
Edoardo Dallanave
}
\note{
}


\seealso{
}
\examples{
##create N-S strained directions from the reference Ardo_Geo_PmagDirs distribution and plots both original and synthetically strained directions on the same equal area

M = matrix_maker(Fol=1,Lin=1.8,v1d=0,v1i=0,v2d=90,v2i=0,v3d=0,v3i=90)
ArdoST = strain_DI(Ardo_Geo_PmagDirs, M)
plot_DI(Ardo_PmagDirs)
plot_DI(bed_DI(ArdoST),on_plot=T,col_d="red",col_u="pink",symbol="d")
}
