% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{T12_GAPWP}
\alias{T12_GAPWP}
\title{Global apparent polar wander path T12}
\format{
## `T12_GAPWP`
A data frame with 33 rows and 16 columns:
\describe{
  \item{Age}{age of the moving window}
  \item{lat}{pole latitude}
  \item{long}{pole longitude; the different coordinate reference systems are AF= south africa
  Nam= North America; EU= Stable Europe; In= India; Amz= Amazonia; Aus= Australia; Eant= East Antarctica}
  ...
}
}
\source{
<https://www.sciencedirect.com/science/article/abs/pii/S0012825212000797>
}
\usage{
T12_GAPWP
}
\description{
Global synthetic apparent polar wander path of Torsvik et al., 2012
(Torsvik,T.H., R. Van der Voo, U. Preeden, C. Mac Niocaill, B. Steinberger, P. V. Doubrovine, D. J. J. van Hinsbergen, M. Domeier, C. Gaina, E. Tohver, J. G. Meert, P. J. a. McCausland, L. R. M. Cocks, Phanerozoic polar wander, palaeogeography and dynamics. Earth-Science Rev. 114, 325–368 (2012).)
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
