% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{V23_GAPWP}
\alias{V23_GAPWP}
\title{Global apparent polar wander path V23}
\format{
## `V23_GAPWP`
A data frame with 33 rows and 20 columns:
\describe{
  \item{Age}{age of the moving window}
  \item{lat}{pole latitude}
  \item{long}{pole longitude; the different coordinate reference systems are SAF= south Africa
  NAm= North America; SAm= South America; EU= Stable Europe; In= India; Aus= Australia;
  Ant= Antarctica; Pac= Pacific; Ib= Iberia}
  ...
}
}
\source{
<https://www.sciencedirect.com/science/article/pii/S0012825223002362>
}
\usage{
V23_GAPWP
}
\description{
Global synthetic apparent polar wander path of Vaes et al., 2023
(Vaes, B., D.J.J. van Hinsbergen, S.H.A. van de Lagemaat, E. van der Wiel, N. Lom, E. Advokaat, L.M. Boschman, L.C. Gallo, A. Greve, C. Guilmette, A global apparent polar wander path for the last 320 Ma calculated from site-level paleomagnetic data. Earth-Science Rev. 245, 1–35 (2023).)
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
