% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{world_coastline}
\alias{world_coastline}
\title{world coastline}
\format{
## `world_coastline`
A data frame with 5261 rows and 2 columns:
\describe{
  \item{lon}{point longitude}
  \item{lat}{point latitude}
  ...
}
}
\usage{
world_coastline
}
\description{
world coastline
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
