% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{world_coastline_180}
\alias{world_coastline_180}
\title{world coastline_180}
\format{
## `world_coastline_180`
A data frame with 5261 rows and 2 columns:
\describe{
  \item{lon}{point longitude adapted for map centerd at 180 degree longitude}
  \item{lat}{point latitude}
  ...
}
}
\source{
https://www.naturalearthdata.com/downloads/110m-physical-vectors/
}
\usage{
world_coastline_180
}
\description{
Detail= 1:110,000,000
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
