% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Ardo_PmagDiR}
\alias{Ardo_PmagDiR}
\title{Ardo section paleomagnetic directions}
\format{
## `Ardo_PmagDiR`
A data frame with 239 rows and 3 columns:
\describe{
  \item{B.Dec}{Paleomagnetic declination}
  \item{B.Inc}{Paleomagnetic inclination}
  \item{Position}{Stratigraphic position}

  ...
}
}
\source{
<https://www.sciencedirect.com/science/article/abs/pii/S0031018212002027>
}
\usage{
Ardo_PmagDiR
}
\description{
Paleomagnetic declination and incliantion, in tilt-correted coordinates, and stratigraphic position from the Paleogene South Ardo section of the Venetian Southern Alps (Italy)
}
\author{
Edoardo Dallanave
}
\keyword{datasets}
